% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AGEB_inegi_2019_a}
\alias{AGEB_inegi_2019_a}
\title{Polygons of AGEBs from México of the National Institute of Geography and Statistics 2019}
\format{
A sf object with 80202 rows and 10 variables:
\describe{
  \item{OBJECTID}{the ids.}
  \item{CVEGEO}{the string key. The key included the CVE_ENT, CVE_MUN, CVE_LOC and CVE_AGEB}
  \item{CVE_ENT}{is the string key for each state.}
  \item{CVE_MUN}{is the string key to the Municipality.}
  \item{CVE_LOC}{is string the key of locality.}
  \item{CVE_AGEB}{is the string key of AGEB.}
  \item{Ambito}{is a binary variable with two values, Urbana and Rural. The dataset is inly AGEB Urban}
  \item{Shape_Leng}{is the length of polygon}
  \item{geometry}{simple feature.}
  ...
}
}
\source{
\url{https://www.inegi.org.mx/temas/mg/#Descargas}
}
\usage{
AGEB_inegi_2019_a
}
\description{
A spatial dataset containing of polygon data of AGEBs from México
}
\details{
AGEB (Areas GeoEstadisticas Basicas) is stand for basic geostatistical areas. This datasets were published in september 2019.
}
\keyword{datasets}
