% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AGEE_inegi19_mx}
\alias{AGEE_inegi19_mx}
\title{Polygons of AGEEs from México of the National Institute of Geography and Statistics 2019
A spatial dataset containing of polygon data of AGEE from México}
\format{
A sf object with 32 rows and 4 variables:
\describe{
  \item{CVEGEO}{the numeric key. The key included the CVE_ENT.}
  \item{CVE_ENT}{is the numeric key for each state.}
  \item{NOMGEO}{is a string value. Its the name of state.}
  \item{geometry}{simple feature.}
  ...
}
}
\source{
\url{https://www.inegi.org.mx/temas/mg/#Descargas}
}
\usage{
AGEE_inegi19_mx
}
\description{
Polygons of AGEEs from México of the National Institute of Geography and Statistics 2019
A spatial dataset containing of polygon data of AGEE from México
}
\details{
AGEE (Areas GeoEstadisticas Estatales) is stand for State Geostatistical Areas. This datasets were published in september 2019.
}
\keyword{datasets}
