% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blocks_ine20_mx_c}
\alias{blocks_ine20_mx_c}
\title{Polygons of blocks from México of the National Electoral Institute 2020}
\format{
A sf object with 2278337 rows and 19 variables:
\describe{
  \item{gid}{the ids.}
  \item{id}{the ids.}
  \item{entidad}{is the numeric key of state.}
  \item{distrito}{is the numeric key of federal district.}
  \item{municipio}{is the numeric key to the Municipality}
  \item{seccion}{is the numeric key to the Sectors}
  \item{localidad}{is numeric the key of locality}
  \item{manzana}{is the numeric key of blocks}
  \item{geometry}{simple feature.}
  ...
}
}
\source{
\url{https://pautas.ine.mx/transparencia/mapas/}
}
\usage{
blocks_ine20_mx_c
}
\description{
A spatial dataset containing of polygon data of blocks from México
}
\details{
This datasets were publiched in october 2020.
}
\keyword{datasets}
