% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blocks_ine20_mx_centroid}
\alias{blocks_ine20_mx_centroid}
\title{Blocks centroids from México of the National Electoral Institute 2020
A dataset containing of centroid data of blocks from México}
\format{
A dataframe with 1556070 rows and 4 variables:
\describe{
  \item{cve_ent}{the numeric key state}
  \item{cve_geo}{is the numeric key for each block centroid.The key included the key of state, municipality, locality, sector and blocks.}
  \item{x}{is the longitud}
  \item{y}{is the latitud}
  ...
}
}
\source{
[INE](https://pautas.ine.mx/transparencia/mapas/)
}
\usage{
data(blocks_ine20_mx_centroid); rgeomex::blocks_ine20_mx_centroid
}
\description{
Blocks centroids from México of the National Electoral Institute 2020
A dataset containing of centroid data of blocks from México
}
\keyword{datasets}
