% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ageb.R
\name{extract_ageb}
\alias{extract_ageb}
\title{extract_ageb}
\usage{
extract_ageb(locality, cve_geo)
}
\arguments{
\item{locality}{locality target.}

\item{cve_geo}{is a string of the state id.}
}
\value{
a list with two sf objects: a locality limit and the agebs of locality target
}
\description{
This function extract the AGEB of the locality.
}
\examples{
1+1
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}.
}
