% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_blocks.R
\name{extract_blocks}
\alias{extract_blocks}
\title{extract_blocks}
\usage{
extract_blocks(locality, cve_edo)
}
\arguments{
\item{locality}{locality target.}

\item{cve_edo}{is a string of the state id.}
}
\value{
a list with two sf objects: a locality limit and the blocks of locality target
}
\description{
This function extract the blocks of the locality.
}
\examples{
x <- rgeomex::extract_blocks(cve_edo = "31", locality = "Merida")
mapview::mapview(x$block)
mapview::mapview(x$locality)
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}.
}
