% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_locality.R
\name{extract_locality}
\alias{extract_locality}
\title{extract the locality}
\usage{
extract_locality(cve_edo, locality)
}
\arguments{
\item{cve_edo}{is the id of state.}

\item{locality}{is the target locality.}
}
\value{
a sf object of the locality boundary.
}
\description{
extract the locality
}
\examples{
rgeomex::extract_locality(cve_edo = "31", locality = "Merida")
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}.
}
