% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_sectores.R
\name{extract_sectores}
\alias{extract_sectores}
\title{extract_sectores}
\usage{
extract_sectores(locality, cve_edo)
}
\arguments{
\item{locality}{locality target.}

\item{cve_edo}{is a string of the state id.}
}
\value{
a list with two sf objects: a boundary of locality and the sectores of locality target
}
\description{
This function extract the sectores of the locality.
}
\examples{
x <- rgeomex::extract_sectores(cve_edo = "31", locality = "Merida")
mapview::mapview(x$sectores)
mapview::mapview(x$locality)
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}.
}
