% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hrl_mx}
\alias{hrl_mx}
\title{hrl_mx}
\format{
A sf object with 150 rows and 8 variables:
\describe{
  \item{cvegeo}{the ids.}
  \item{loc}{name of the localitie}
  \item{ambito}{type of localities}
  \item{geometry}{geometry column}
  ...
}
}
\source{
\url{https://geosis.mx/Aplicaciones/EntomologiayControl/Home/IndexNacional}
}
\usage{
hrl_mx
}
\description{
High risk localities of Mexico
}
\details{
This database was built using the SIMN annual goals dataset
}
\keyword{datasets}
