% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iter_2020_loc}
\alias{iter_2020_loc}
\title{Iter 2020}
\format{
A sf object with 189432 rows and 10 variables:
\describe{
  \item{entidad}{the ids.}
  \item{nom_ent}{the ids.}
  \item{mun}{is the numeric key of state.}
  \item{nom_mun}{is the numeric key of federal district.}
  \item{nom_loc}{is the numeric key of federal district.}
  \item{loc}{is the numeric key to the Municipality}
  \item{longitud}{is the numeric key to the Sectors}
  \item{latitud}{is numeric the key of tipo}
  \item{altitud}{is the simple feature}
  \item{pobtot}{}
  ...
}
}
\source{
\url{https://www.inegi.org.mx/programas/ccpv/2020/#datos_abiertos}
}
\usage{
iter_2020_loc
}
\description{
A spatial dataset containing of the National Institute of Geography and Statistics INEGI ITER 2020
}
\details{
This datasets were published in July 26 2021.
}
\keyword{datasets}
