% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{loc_inegi19_mx}
\alias{loc_inegi19_mx}
\title{Polygons of localities from México of the National Institute of Geography and Statistics 2019
A spatial dataset containing of polygon data of localities from México}
\format{
A sf object with 49981 rows and 7 variables:
\describe{
  \item{CVEGEO}{the numeric key. The key included the CVE_ENT and the CVE_MUN.}
  \item{CVE_ENT}{is the numeric key for each state.}
  \item{CVE_MUN}{is the string key to the Municipality.}
  \item{CVE_LOC}{is the string key of locality.}
  \item{NOMGEO}{is a string value. Its the name of locality.}
  \item{Ambito}{is a binary variable with two values, Urbana and Rural.}
  \item{geometry}{simple feature.}
  ...
}
}
\source{
\url{https://www.inegi.org.mx/temas/mg/#Descargas}
}
\usage{
loc_inegi19_mx
}
\description{
Polygons of localities from México of the National Institute of Geography and Statistics 2019
A spatial dataset containing of polygon data of localities from México
}
\details{
This datasets were published in september 2019.
}
\keyword{datasets}
