% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pua_mx}
\alias{pua_mx}
\title{pua_mx}
\format{
A sf object with 111 rows and 3 variables:
\describe{
  \item{id}{the ids.}
  \item{loc}{name of the priority urban area}
  \item{geometry}{geometry column}
  ...
}
}
\source{
\url{https://geosis.mx/Aplicaciones/EntomologiayControl/Home/IndexNacional}
}
\usage{
pua_mx
}
\description{
Priority Urban Areas of Mexico
}
\details{
This database was built using data from high-risk locations, called hrl_mx in rgeomex
}
\keyword{datasets}
