% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sectores_ine20_mx_b}
\alias{sectores_ine20_mx_b}
\title{Polygons of sectores from México of the National Electoral Institute 2020}
\format{
A sf object with 27428 rows and 7 variables:
\describe{
  \item{gid}{the ids.}
  \item{id}{the ids.}
  \item{entidad}{is the numeric key of state.}
  \item{distrito}{is the numeric key of federal district.}
  \item{distrito_l}{is the numeric key of federal district.}
  \item{municipio}{is the numeric key to the Municipality}
  \item{seccion}{is the numeric key to the Sectors}
  \item{tipo}{is numeric the key of tipo}
  \item{geometry1_}{is the simple feature}
  \item{ln}{}
  \item{pad}{}
  \item{geometry}{simple feature.}
  ...
}
}
\source{
\url{https://pautas.ine.mx/transparencia/mapas/}
}
\usage{
sectores_ine20_mx_b
}
\description{
A spatial dataset containing of polygon data of sectores from México
}
\details{
This datasets were published in october 2023.
}
\keyword{datasets}
