% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ua_mx}
\alias{ua_mx}
\title{ua_mx}
\format{
A sf object with 658 rows and 2 variables:
\describe{
  \item{id}{the ids.}
  \item{geometry}{geometry column}
  ...
}
}
\source{
\url{https://dataforgood.facebook.com/dfg/tools/globalurbanareas}
}
\usage{
ua_mx
}
\description{
Urban Areas of Mexico
}
\details{
La base de las area urbana de mexico derivada de la base de las areas global urbanas de meta data for goof of facebook
}
\keyword{datasets}
