% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AGEM_inegi19_mx}
\alias{AGEM_inegi19_mx}
\title{Polygons of AGEMs from México of the National Institute of Geography and Statistics 2019
A spatial dataset containing of polygon data of AGEM from México}
\format{
A sf object with 2465 rows and 5 variables:
\describe{
  \item{CVEGEO}{the numeric key. The key included the CVE_ENT and the CVE_MUN.}
  \item{CVE_ENT}{is the numeric key for each state.}
  \item{CVE_MUN}{is the string key to the Municipality.}
  \item{NOMGEO}{is a string value. Its the name of municipality.}
  \item{geometry}{simple feature.}
  ...
}
}
\source{
\url{https://www.inegi.org.mx/temas/mg/#Descargas}
}
\usage{
AGEM_inegi19_mx
}
\description{
Polygons of AGEMs from México of the National Institute of Geography and Statistics 2019
A spatial dataset containing of polygon data of AGEM from México
}
\details{
AGEM (Areas GeoEstadisticas Municipales) is stand for Municipality Geostatistical Areas. This datasets were published in september 2019.
}
\keyword{datasets}
