% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ageb.R
\name{extract_ageb}
\alias{extract_ageb}
\title{extract_ageb}
\usage{
extract_ageb(locality, cve_edo)
}
\arguments{
\item{locality}{locality target.}

\item{cve_edo}{is a string of the state id.}
}
\value{
a list with two sf objects: a locality limit and the agebs of locality target
}
\description{
This function extract the AGEB of the locality.
}
\examples{
x <- rgeomex::extract_ageb(cve_edo = "31", locality = "Merida")
mapview::mapview(x$ageb)
mapview::mapview(x$locality)
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}.
}
