% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgeomex-package.R
\docType{package}
\name{rgeomex-package}
\alias{rgeomex}
\alias{rgeomex-package}
\title{rgeomex: The rgeomex provides the polygon data of INEGI 2019 and INE 2020}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

the rgeomex package provides spatial dataset (areal data) available and published homepage of INE and INEGI.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fdzul/rgeomex}
  \item \url{https://fdzul.github.io/rgeomex/}
  \item Report bugs at \url{https://github.com/fdzul/rgeomex/issues}
}

}
\author{
\strong{Maintainer}: Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com} (\href{https://orcid.org/0000-0002-2085-9141}{ORCID})

Other contributors:
\itemize{
  \item Fabian Correa-Morales \email{fabiancorrea@msn.com} (\href{https://orcid.org/0000-0002-6193-1242}{ORCID}) [contributor]
}

}
\keyword{internal}
